Add unit test code coverage reporting (#1399)
* Allow board to override ATOMIC() macro

* Override native GCC _countof()

* Add cast for native build

* Rename SERIAL_BUFFER_SIZE - clashes with an ArduinoFake define

* Use delayMicroseconds() for NATIVE_BOARD

* Native unit tests with code coverage using ArduinoFake

* Incorporate old native test into code coverage

* Add code coverage Github action

* gcovr action

* Coverage action: enable all unit tests

* Use a 1ms thread based timer to simulate hardware timers

* Upload code coverage report to CodeCov.io

* Conditional compile FRAM EEPROM code (this will exclude it from code coverage)

* Remove folder exclusions for conditionally compiled code

* Add CodeCov badge

+++ speeduino/scheduler.cpp
@@ -307,7 +307,7 @@ static inline void setScheduleRunning(Schedule &schedule, uint32_t delay, uint16
 
 void setSchedule(Schedule &schedule, uint32_t delay, uint16_t duration, bool allowQueuedSchedule)
 {
-  if((delay>0U) && (delay < MAX_TIMER_PERIOD) && (duration > 0U))
+  if((delay>0U) && (delay < (uint32_t)MAX_TIMER_PERIOD) && (duration > 0U))
   {
     ATOMIC() 
     {
